extract mongo-derivation
First step of https://github.com/sphereio/sphere-scala-libs/issues/174
