+++ mongo/mongo-core/dependencies.sbt
