+++ mongo/mongo-core/src/main/scala/io/sphere/mongo/catsinstances/package.scala
