+++ mongo/mongo-core/src/main/scala/io/sphere/mongo/format/MongoFormat.scala
