+++ mongo/mongo-core/src/main/scala/io/sphere/mongo/format/package.scala
