+++ mongo/mongo-derivation/src/main/java/io/sphere/mongo/generic/annotations/MongoIgnore.java
