+++ mongo/mongo-derivation/src/test/scala/io/sphere/mongo/MongoUtils.scala
@@ -1,0 +1,9 @@
+package io.sphere.mongo
+import com.mongodb.BasicDBObject
+
+object MongoUtils {
+
+  def dbObj(pairs: (String, Any)*) =
+    pairs.foldLeft(new BasicDBObject){case (obj, (key, value)) => obj.append(key, value)}
+
+}
